+++ source/include (local)/managed_key_value_array_macros.h
@@ -1,4 +1,5 @@
 #include <stdbool.h>
+#include <stddef.h>
 
 #define _IoTLib_declare_managed_key_value_array_insert_function(managedKeyValueArrayTypeName, keyType, valueType) \
 	void managedKeyValueArrayTypeName##_insert(struct managedKeyValueArrayTypeName* keyValueArray, const keyType key, valueType value)
@@ -18,7 +19,7 @@
 #define _IoTLib_define_managed_key_value_array_get_function(managedKeyValueArrayTypeName, keyType, valueType) \
 	valueType managedKeyValueArrayTypeName##_get(const struct managedKeyValueArrayTypeName* keyValueArray, const keyType key) \
 	{ \
-		int index = keyType##_get_key_index(keyValueArray->keys, keyValueArray->length, key); \
+		size_t index = keyType##_get_key_index(keyValueArray->keys, keyValueArray->length, key); \
 		IoTLib_managed_key_value_array_getFunc_validate_key_index(index); \
 		return keyValueArray->values[index]; \
 	}
@@ -29,7 +30,7 @@
 #define _IoTLib_define_managed_key_value_array_try_get_function(managedKeyValueArrayTypeName, keyType, valueType) \
 	bool managedKeyValueArrayTypeName##_tryGet(const struct managedKeyValueArrayTypeName* keyValueArray, const keyType key, valueType* valueTypeOut) \
 	{ \
-		int index = keyType##_get_key_index(keyValueArray->keys, keyValueArray->length, key); \
+		size_t index = keyType##_get_key_index(keyValueArray->keys, keyValueArray->length, key); \
 		if (IoTLib_managed_key_value_array_index_is_valid_index(index)) \
 		{ \
 			*valueTypeOut = keyValueArray->values[index]; \
@@ -44,7 +45,7 @@
 #define _IoTLib_define_managed_key_value_array_contains_key_function(managedKeyValueArrayTypeName, keyType) \
 	bool managedKeyValueArrayTypeName##_containsKey(const struct managedKeyValueArrayTypeName* keyValueArray, const keyType key) \
 	{ \
-		int index = keyType##_get_key_index(keyValueArray->keys, keyValueArray->length, key); \
+		size_t index = keyType##_get_key_index(keyValueArray->keys, keyValueArray->length, key); \
 		return IoTLib_managed_key_value_array_index_is_valid_index(index); \
 	}
 
@@ -63,7 +64,7 @@
 
 
 #define IoTLib_define_functions_for_key_type(keyType, getKeyIndexFunc) \
-	int keyType##_get_key_index(const keyType* keyArray, const int arrayLength, const keyType keyToSearchFor) \
+	size_t keyType##_get_key_index(const keyType* keyArray, const size_t arrayLength, const keyType keyToSearchFor) \
 	{ \
 		return getKeyIndexFunc(keyArray, arrayLength, keyToSearchFor); \
 	}
@@ -73,8 +74,8 @@
 	{ \
 		keyType* keys; \
 		valueType* values; \
-		int length; \
-		int capacity; \
+		size_t length; \
+		size_t capacity; \
 	}; \
 	_IoTLib_declare_managed_key_value_array_functions(managedKeyValueArrayTypeName, keyType, valueType)
